Added path base. Added path composition observer. Added path renderer. Added misc changes. NO SERIALIZATION !

+++ core/src/main/java/com/draniksoft/ome/support/ui/util/CompObViewIds.java
@@ -1,5 +1,5 @@
 package com.draniksoft.ome.support.ui.util;
 
 public class CompObViewIds {
-    public static final int VIEW_INSPECTOR = 1;
+    public static final short VIEW_INSPECTOR = 1;
 }
